applied feedback from ciarands
